requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asttokens
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama
comm
debugpy
decorator
exceptiongroup
executing
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
idna==3.7
importlib_metadata
ipykernel
ipython
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jupyter_client
jupyter_core
MarkupSafe==2.1.5
matplotlib-inline
multidict==6.0.5
nest_asyncio
numpy==2.0.1
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging
parso
pdf2image==1.17.0
pickleshare
pillow==10.4.0
platformdirs
prompt_toolkit
psutil
pure_eval
pycparser==2.22
Pygments
PyJWT==2.8.0
pyngrok==7.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil
PyYAML==6.0.1
gunicorn==22.0.0
pyzmq
requests==2.32.3
selenium==4.23.1
setuptools==71.0.4
six
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data
tesseract==0.1.3
tornado
traitlets
trio==0.26.0
trio-websocket==0.11.1
twilio==9.2.3
typing_extensions
urllib3==2.2.2
wcwidth
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
yarl==1.9.4
zipp
